# docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: localhost:25565
info:
  description: An API used to do banking operations such as user register, withdrawl, deposit, access control, internal and external accounts transfers, account blocking and more.
  version: "1.0.0"
  title: "Simple Banking API"
  contact:
    email: julio.kobayashi@ebury.com
  license:
    name: Apache 2.0
    url: "[http://www.apache.org/licenses/LICENSE-2.0.html](http://www.apache.org/licenses/LICENSE-2.0.html)"
...
paths:
  /users:
    post:
      tags:
        - users
      summary: creates user with info passed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        "201":
          description: user created sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Bad input parameter
        "500":
          description: Internal Server Error
  /users/auth:
    put:
      tags:
        - users
      summary: authenticates a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                password:
                  type: string
      responses:
        "202":
          description: User Authorized
          headers:
            Authorization:
              schema:
                type: string
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /users/{user_id}:
    get:
      tags:
        - users
      summary: return data of user with ID userID
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: user returned succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
    delete:
      tags:
        - users
      summary: delete user of userID
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
    put:
      tags:
        - users
      summary: update a given piece of information of a certain user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        "200":
          description: User updated sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /users/{user_id}/report:
    get:
      tags:
        - users
      summary: returns a report of the user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserReport'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /users/{user_id}/block:
    put:
      tags:
        - users
      summary: block a user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User Blocked
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /users/{user_id}/unblock:
    put:
      tags:
        - users
      summary: unblock a user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User Unblocked
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
...
  /withdrawals:
    post:
      tags:
        - withdrawals
      summary: creates a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Withdrawal'
      responses:
        "201":
          description: Withdrawal created sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Withdrawal'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /withdrawals/{withdrawal_id}:
    get:
      tags:
        - withdrawals
      summary: return data of withdrawal with ID withdrawalID
      parameters:
        - in: path
          name: withdrawal_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: withdrawal returned sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Withdrawal'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
    delete:
      tags:
        - withdrawals
      summary: delete withdrawal of withdrawalID
      parameters:
        - in: path
          name: withdrawal_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Withdrawal Deleted!
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
...
components:
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
        name:
          type: string
        document:
          type: string
        password:
          type: string
        register_date:
          type: string
          format: date-time
        status:
          type: boolean
    Client:
      type: object
      properties:
        client_id:
          type: string
        user_id:
          type: string
        name:
          type: string
        document:
          type: string
        register_date:
          type: string
          format: date-time
        status:
          type: boolean
...
    Withdrawal:
      type: object
      properties:
        withdrawal_id:
          type: string
        account_id:
          type: string
        client_id:
          type: string
        agency_id:
          type: integer
          format: uint32
        withdrawal:
          type: number
          format: float64
        withdrawal_date:
          type: string
          format: date-time
        status:
          type: boolean
...
    UserReport:
      type: object
      properties:
        user_id:
          type: string
        name:
          type: string
        document:
          type: string
        register_date:
          type: string
          format: date-time
        status:
          type: boolean
        clients:
          type: array
          items:
            $ref: '#/components/schemas/Client'
        report_date:
          type: string
          format: date-time
...